+++ src/colruyt/pcrsejb/service/dl/competence/CompetenceService.java
@@ -5,7 +5,7 @@ import colruyt.pcrsejb.service.dl.*;
 
 import java.util.*;
 
-public interface CompetenceService extends GenericCrudService<Competence,Long> {
+public interface CompetenceService extends GenericCrudService<Competence,Integer> {
 	
 	public Collection<FunctionCompetence> findAllFunctionCompetences();
 	public Collection<CraftCompetence> findAllCraftCompetences();
