+++ src/colruyt/pcrsejb/service/dl/function/FunctionService.java
@@ -3,7 +3,7 @@ package colruyt.pcrsejb.service.dl.function;
 import colruyt.pcrsejb.entity.function.Function;
 import colruyt.pcrsejb.service.dl.GenericCrudService;
 
-public interface FunctionService extends GenericCrudService<Function,Long> {
+public interface FunctionService extends GenericCrudService<Function,Integer> {
 
 
 }
