+++ src/colruyt/pcrsejb/service/dl/survey/SurveySetService.java
@@ -3,5 +3,5 @@ package colruyt.pcrsejb.service.dl.survey;
 import colruyt.pcrsejb.entity.survey.SurveySet;
 import colruyt.pcrsejb.service.dl.GenericCrudService;
 
-public interface SurveySetService extends GenericCrudService<SurveySet,Long> {
+public interface SurveySetService extends GenericCrudService<SurveySet,Integer> {
 }
